Removd unused variables and #include <stdlib.h>
2002-11-07  Jeffrey Stedfast  <fejj@ximian.com>

	* iconv-detect.c: Removd unused variables and #include <stdlib.h>

	* gmime/gmime-multipart.c
	(g_mime_multipart_get_subpart_from_content_id): Init part to NULL.

	* gmime/gmime-stream-cat.c (stream_write): Init n to -1 each time
	through the outer loop.

	* gmime/gmime-filter-strip.c: #include <string.h> for memcpy.

	* gmime/gmime-filter-html.c (html_convert): Make sure to
	initialise depth to 0 each loop so we don't use it uninitialised.

+++ gmime/gmime-charset.c
@@ -311,7 +311,7 @@ g_mime_charset_name (const char *charset)
 void
 g_mime_charset_init (GMimeCharset *charset)
 {
-	charset->mask = ~0;
+	charset->mask = (unsigned int) ~0;
 	charset->level = 0;
 }
